+++ app/models/generic_event.rb
@@ -1,4 +1,5 @@
 class GenericEvent < ApplicationRecord
+  DETAILS_ATTRIBUTES = %w[].freeze
   FEED_ATTRIBUTES = %w[
     id
     notes