+++ app/models/generic_event/journey_handover_to_destination.rb
@@ -1,5 +1,9 @@
 class GenericEvent
   class JourneyHandoverToDestination < GenericEvent
+    DETAILS_ATTRIBUTES = %w[
+      supplier_personnel_id
+    ].freeze
+
     include JourneyEventValidations
 
     def supplier_personnel_id