+++ app/models/generic_event/move_notify_premises_of_eta.rb
@@ -1,5 +1,9 @@
 class GenericEvent
   class MoveNotifyPremisesOfEta < GenericEvent
+    DETAILS_ATTRIBUTES = %w[
+      expected_at
+    ].freeze
+
     include MoveEventValidations
 
     validates :expected_at, presence: true