+++ app/models/generic_event/move_operation_tornado.rb
@@ -1,5 +1,10 @@
 class GenericEvent
   class MoveOperationTornado < GenericEvent
+    DETAILS_ATTRIBUTES = %w[
+      authorised_at
+      authorised_by
+    ].freeze
+
     include MoveEventValidations
     include AuthoriserValidations
   end